1.21

+++ common/src/main/java/com/daqem/jobsplus/player/job/exp/ExpCollector.java
@@ -4,14 +4,14 @@ import com.daqem.jobsplus.JobsPlus;
 
 public class ExpCollector {
 
-    private int jobExp = 0;
+    private double jobExp = 0;
 
-    public void addExp(int exp) {
+    public void addExp(double exp) {
         JobsPlus.debug("Adding {} experience to the job.", exp);
         jobExp += exp;
     }
 
-    public int getExp() {
+    public double getExp() {
         return jobExp;
     }
